move demo plugins from UIUGens.cpp/.mm to a dedicated file DemoUGens.cpp
