+++ server/plugins/UIUGens.cpp
@@ -259,130 +259,6 @@ void MouseButton_Ctor(MouseInputUGen* unit) {
     MouseButton_next(unit, 1);
 }
 
-//////////////////////////////////////////////////////////////////////////////////////////////////
-//////////////////////////////////////////////////////////////////////////////////////////////////
-//////////////////////////////////////////////////////////////////////////////////////////////////
-
-// example of implementing a plug in command with async execution.
-
-struct MyPluginData // data for the global instance of the plugin
-{
-    float a, b;
-};
-
-struct MyCmdData // data for each command
-{
-    MyPluginData* myPlugin;
-    float x, y;
-    char* name;
-};
-
-MyPluginData gMyPlugin; // global
-
-bool cmdStage2(World* world, void* inUserData) {
-    // user data is the command.
-    MyCmdData* myCmdData = (MyCmdData*)inUserData;
-
-    // just print out the values
-    Print("cmdStage2 a %g  b %g  x %g  y %g  name %s\n", myCmdData->myPlugin->a, myCmdData->myPlugin->b, myCmdData->x,
-          myCmdData->y, myCmdData->name);
-
-    return true;
-}
-
-bool cmdStage3(World* world, void* inUserData) {
-    // user data is the command.
-    MyCmdData* myCmdData = (MyCmdData*)inUserData;
-
-    // just print out the values
-    Print("cmdStage3 a %g  b %g  x %g  y %g  name %s\n", myCmdData->myPlugin->a, myCmdData->myPlugin->b, myCmdData->x,
-          myCmdData->y, myCmdData->name);
-
-    // scsynth will perform completion message after this returns
-    return true;
-}
-
-bool cmdStage4(World* world, void* inUserData) {
-    // user data is the command.
-    MyCmdData* myCmdData = (MyCmdData*)inUserData;
-
-    // just print out the values
-    Print("cmdStage4 a %g  b %g  x %g  y %g  name %s\n", myCmdData->myPlugin->a, myCmdData->myPlugin->b, myCmdData->x,
-          myCmdData->y, myCmdData->name);
-
-    // scsynth will send /done after this returns
-    return true;
-}
-
-void cmdCleanup(World* world, void* inUserData) {
-    // user data is the command.
-    MyCmdData* myCmdData = (MyCmdData*)inUserData;
-
-    Print("cmdCleanup a %g  b %g  x %g  y %g  name %s\n", myCmdData->myPlugin->a, myCmdData->myPlugin->b, myCmdData->x,
-          myCmdData->y, myCmdData->name);
-
-    RTFree(world, myCmdData->name); // free the string
-    RTFree(world, myCmdData); // free command data
-    // scsynth will delete the completion message for you.
-}
-
-void cmdDemoFunc(World* inWorld, void* inUserData, struct sc_msg_iter* args, void* replyAddr) {
-    Print("->cmdDemoFunc %p\n", inUserData);
-
-    // user data is the plug-in's user data.
-    MyPluginData* thePlugInData = (MyPluginData*)inUserData;
-
-    // allocate command data, free it in cmdCleanup.
-    MyCmdData* myCmdData = (MyCmdData*)RTAlloc(inWorld, sizeof(MyCmdData));
-    myCmdData->myPlugin = thePlugInData;
-
-    // ..get data from args..
-    myCmdData->x = 0.;
-    myCmdData->y = 0.;
-    myCmdData->name = 0;
-
-    // float arguments
-    myCmdData->x = args->getf();
-    myCmdData->y = args->getf();
-
-    // how to pass a string argument:
-    const char* name = args->gets(); // get the string argument
-    if (name) {
-        myCmdData->name = (char*)RTAlloc(inWorld, strlen(name) + 1); // allocate space, free it in cmdCleanup.
-        strcpy(myCmdData->name, name); // copy the string
-    }
-
-    // how to pass a completion message
-    int msgSize = args->getbsize();
-    char* msgData = 0;
-    if (msgSize) {
-        // allocate space for completion message
-        // scsynth will delete the completion message for you.
-        msgData = (char*)RTAlloc(inWorld, msgSize);
-        args->getb(msgData, msgSize); // copy completion message.
-    }
-
-    DoAsynchronousCommand(inWorld, replyAddr, "cmdDemoFunc", (void*)myCmdData, (AsyncStageFn)cmdStage2,
-                          (AsyncStageFn)cmdStage3, (AsyncStageFn)cmdStage4, cmdCleanup, msgSize, msgData);
-
-    Print("<-cmdDemoFunc\n");
-}
-
-/*
- * to test the above, send the server these commands:
- *
- *
- * SynthDef(\sine, { Out.ar(0, SinOsc.ar(800,0,0.2)) }).load(s);
- * s.sendMsg(\cmd, \pluginCmdDemo, 7, 9, \mno, [\s_new, \sine, 900, 0, 0]);
- * s.sendMsg(\n_free, 900);
- * s.sendMsg(\cmd, \pluginCmdDemo, 7, 9, \mno);
- * s.sendMsg(\cmd, \pluginCmdDemo, 7, 9);
- * s.sendMsg(\cmd, \pluginCmdDemo, 7);
- * s.sendMsg(\cmd, \pluginCmdDemo);
- *
- */
-
-
 SC_Thread uiListenThread;
 
 PluginLoad(UIUGens) {
@@ -396,11 +272,6 @@ PluginLoad(UIUGens) {
     DefineUnit("MouseX", sizeof(MouseInputUGen), (UnitCtorFunc)&MouseX_Ctor, 0, 0);
     DefineUnit("MouseY", sizeof(MouseInputUGen), (UnitCtorFunc)&MouseY_Ctor, 0, 0);
     DefineUnit("MouseButton", sizeof(MouseInputUGen), (UnitCtorFunc)&MouseButton_Ctor, 0, 0);
-
-    // define a plugin command - example code
-    gMyPlugin.a = 1.2f;
-    gMyPlugin.b = 3.4f;
-    DefinePlugInCmd("pluginCmdDemo", cmdDemoFunc, (void*)&gMyPlugin);
 }
